+++ src/app/domain-of-influences/domain-of-influence-voting-card-data-edit/domain-of-influence-voting-card-data-edit.component.ts
@@ -21,7 +21,6 @@ import { VotingCardColor } from '@abraxas/voting-basis-service-proto/grpc/shared
 @Component({
   selector: 'app-domain-of-influence-voting-card-data-edit',
   templateUrl: './domain-of-influence-voting-card-data-edit.component.html',
-  styleUrl: 'domain-of-influence-voting-card-data-edit.component.scss',
   standalone: false,
 })
 export class DomainOfInfluenceVotingCardDataEditComponent implements OnInit {
@@ -111,14 +110,5 @@ export class DomainOfInfluenceVotingCardDataEditComponent implements OnInit {
       );
 
     this.canEditEverything = await this.permissionService.hasPermission(Permissions.DomainOfInfluence.UpdateSameCanton);
-  }
-
-  protected updateStistatExportEnabled(enabled: boolean) {
-    this.domainOfInfluence.stistatExportEnabled = enabled;
-    if (!enabled) {
-      this.domainOfInfluence.stistatExportEaiMessageType = '';
-    }
-
-    this.contentChanged.emit();
   }
 }