+++ src/app/shared/secondary-majority-election-candidates/secondary-majority-election-candidates.component.ts
@@ -49,7 +49,7 @@ export class SecondaryMajorityElectionCandidatesComponent {
   public locked: boolean = false;
 
   @Input()
-  public eVotingApproved: boolean = false;
+  public eVotingApprovedAndInTestingPhase: boolean = false;
 
   @Input()
   public eVotingEverApproved: boolean = false;